chore(ui): refactor query cache
