+++ ui/src/features/project/list/create-project-modal.tsx
@@ -10,9 +10,9 @@ import { z } from 'zod';
 import { YamlEditor } from '@ui/features/common/code-editor/yaml-editor';
 import { FieldContainer } from '@ui/features/common/form/field-container';
 import { ModalComponentProps } from '@ui/features/common/modal/modal-context';
+import { queryCache } from '@ui/features/utils/cache';
 import schema from '@ui/gen/schema/projects.kargo.akuity.io_v1alpha1.json';
 import { createResource } from '@ui/gen/service/v1alpha1/service-KargoService_connectquery';
-import { queryCache } from '@ui/utils/cache';
 import { decodeUint8ArrayYamlManifestToJson } from '@ui/utils/decode-raw-data';
 import { zodValidators } from '@ui/utils/validators';
