+++ ui/src/features/utils/cache/index.ts
@@ -1,0 +1,9 @@
+// cache invalidation source-of-truth
+
+import analysisTemplates from './analysis-templates';
+import project from './project';
+
+export const queryCache = {
+  project,
+  analysisTemplates
+};
